src/main.py
import argparse
import os
import shutil
from tqdm import tqdm
import logging
from src.utils.common import read_yaml, create_directories
import random
import mlflow

STAGE = "MAIN"
create_directories("logs")
with open(os.path.join("logs", "running_logs.log"), 'a') as f:
    f.write("")

logging.basicConfig(
    filename=os.path.join("logs", "running_logs.log"),
    level=logging.INFO,
    format="[%(asctime)s: %(levelname)s: %(module)s]: %(message)s",
    filemode="a"
)
