requirements.txt
### dependency
streamlit
notebook
nltk
sklearn
mlflow
PyYAML
### local packages -
-e . 
